feature and chore: Add footer and scroll-up components
This commit refactors the code to add the footer and scroll-up components to the App.tsx file. The footer component is responsible for displaying social media links and copyright information, while the scroll-up component provides a button to scroll back to the top of the page. The necessary CSS files for both components are also included.

Refactor the Tools component

This commit refactors the Tools component in the Skills folder. It adds the "vite" tool to the list of tools and updates the name of the "Docker" tool to lowercase.

Add useScrollVisibility hook

This commit adds a new hook called useScrollVisibility in the hooks folder. This hook is responsible for adding a CSS class to the scroll-up component when the user scrolls down the page.

Update footer CSS

This commit adds CSS styles to the footer component. It includes styles for the container, title, links, social icons, and copyright text.

Add scroll-up CSS

This commit adds CSS styles to the scroll-up component. It includes styles for the container, icon, and hover effect.

Update dependencies

This commit updates the Node.js and yarn versions in the .tool-versions file. It also updates the dependencies in the package.json and yarn.lock files to add the testing library.

Remove unused Vite logo SVG file and update age in About component

This commit removes the unused Vite logo SVG file from the project. It also updates the age displayed in the About component.

Update meta tags in index.html

This commit updates the meta tags in the index.html file.

Add linters workflow for code linting

This commit adds a linters workflow for code linting. It ensures that the code follows the established linting rules.