tests: Add tests for language change component
Added tests for the language change component to ensure that it renders two buttons for language selection and that the flag for Brazil is selected by default. The tests use the react-testing-library and vitest libraries for testing. This change improves the test coverage and ensures the correct behavior of the component.

refactor: Update image language component

Updated the image language component to use an enum for language types instead of hardcoding the language values. This change improves code readability and maintainability. The component now uses the enum values for the selected language and the onChangeLanguage function. Additionally, the component now renders the correct className for the button based on whether it is selected or not.

refactor: Update i18n component

Updated the i18n component to use the LanguageTypes enum for the selected language and the onChangeLanguage function. This change improves code readability and maintainability. The component now uses the enum values for the selected language and the onChangeLanguage function. The component also imports the LanguageTypes enum from the image-language component.

refactor: Update file paths and extensions for language flags and project images

Updated the file paths and extensions for the language flags and project images in the change-language component. This change ensures that the correct images are displayed and improves the overall organization of the project.

refactor: Update Portuguese translation in home page

Updated the Portuguese translation in the home page component. This change ensures that the translation is accurate and up to date.

refactor: Update Portuguese translation, footer CSS, header link, translation keys, width images to improve the Google Lighthouse report

Updated the Portuguese translation, footer CSS, header link, translation keys, and width images in the project to improve the Google Lighthouse report. This change optimizes the performance and accessibility of the project.